BSIS-961 Move query findComponentByCodeDINAndInventoryStatus to InventoryRepository
