+++ test/repository/InventoryRepositoryTests.java
@@ -2,8 +2,12 @@ package repository;
 
 import static helpers.builders.ComponentBuilder.aComponent;
 import static helpers.builders.DonationBuilder.aDonation;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
 import java.util.List;
+
+import javax.persistence.NoResultException;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -12,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import dto.StockLevelDTO;
 import helpers.builders.ComponentTypeBuilder;
 import helpers.builders.LocationBuilder;
+import model.component.Component;
 import model.componenttype.ComponentType;
 import model.donation.Donation;
 import model.inventory.InventoryStatus;
@@ -129,6 +134,63 @@ public class InventoryRepositoryTests extends ContextDependentTestSuite {
     Assert.assertEquals("Verify blood group", "A+", levels.get(0).getBloodAbo() + levels.get(0).getBloodRh());
     Assert.assertEquals("Verify blood group", "A+", levels.get(1).getBloodAbo() + levels.get(0).getBloodRh());
 
+  }
+  
+  @Test
+  public void testFindComponentByCodeDINAndInventoryStatus_shouldReturnMatchingComponent() {
+    
+    String componentCode = "0011-01";
+    String donationIdentificationNumber = "0000002";
+    
+    Donation donationWithExpectedDonationIdentificationNumber = aDonation()
+        .withDonationIdentificationNumber(donationIdentificationNumber)
+        .buildAndPersist(entityManager);
+    
+    // Excluded by component code
+    aComponent()
+        .withComponentCode("0011-02")
+        .withDonation(donationWithExpectedDonationIdentificationNumber)
+        .buildAndPersist(entityManager);
+    
+    // Excluded by donation identification number
+    aComponent()
+        .withComponentCode(componentCode)
+        .withDonation(aDonation().withDonationIdentificationNumber("1000007").build())
+        .buildAndPersist(entityManager);
+    
+    // Excluded by inventoryStatus = NOT_LABELLED
+    aComponent()
+        .withComponentCode(componentCode)
+        .withDonation(donationWithExpectedDonationIdentificationNumber)
+        .withInventoryStatus(InventoryStatus.NOT_LABELLED)
+        .buildAndPersist(entityManager);
+    
+    // Excluded by inventoryStatus = REMOVED
+    aComponent()
+        .withComponentCode(componentCode)
+        .withDonation(donationWithExpectedDonationIdentificationNumber)
+        .withInventoryStatus(InventoryStatus.REMOVED)
+        .buildAndPersist(entityManager);
+    
+    // Expected
+    Component expectedComponent = aComponent()
+        .withComponentCode(componentCode)
+        .withDonation(donationWithExpectedDonationIdentificationNumber)
+        .withInventoryStatus(InventoryStatus.IN_STOCK)
+        .buildAndPersist(entityManager);
+    
+    // Test
+    Component returnedComponent = inventoryRepository.findComponentByCodeDINAndInventoryStatus(componentCode,
+        donationIdentificationNumber, InventoryStatus.IN_STOCK);
+    
+    // Verify
+    assertThat(returnedComponent, is(expectedComponent));
+  }
+
+  @Test(expected = NoResultException.class)
+  public void testFindNonExistentComponentByCodeDINAndInventoryStatus_shoulThrow() {
+    // Test
+    inventoryRepository.findComponentByCodeDINAndInventoryStatus("0011-01", "0000002", InventoryStatus.IN_STOCK);
   }
 
 }
